+++ packages/frontend/src/components/TableCalculationModals/TableCalculationModal.tsx
@@ -336,7 +336,8 @@ const TableCalculationModal: FC<Props> = ({
                     ? {
                           position: 'absolute',
                           width: '100%',
-                          height: '100%',
+                          height: 'calc(100% - 50px)', // Extracts the height of the header
+                          marginTop: '-60px',
                       }
                     : undefined
             }